Fixed PlaylistRemovePopup bug
YouTube changed the trash icon
from this: https://imgur.com/a/WQE3aD1
to this: https://imgur.com/a/zCoh1cA

When you want to delete a video out of a playlist, you click the trash icon and the script searches the actual "remove from playlist" button that's inside the popup menu. If the script would search for the text: "remove from playlist" this feature would stop working when you'd change your language. So to keep things universal, I search the right vector value for the trash icon instead. These values change as icons get updated.